+++ clients/fides-js/rollup.config.mjs
@@ -3,6 +3,7 @@ import dts from "rollup-plugin-dts";
 import esbuild from "rollup-plugin-esbuild";
 import filesize from "rollup-plugin-filesize";
 import nodeResolve from "@rollup/plugin-node-resolve";
+import css from "rollup-plugin-import-css";
 
 const name = "fides";
 const isDev = process.env.NODE_ENV === "development";
@@ -15,8 +16,11 @@ const GZIP_SIZE_WARN_KB = 15; // log a warning if bundle size exceeds this
 export default [
   {
     input: `src/${name}.ts`,
+    // DEFER: Add aliases for typical react imports (see https://preactjs.com/guide/v10/getting-started/#aliasing-in-rollup)
+    // This will be needed if & when we want to leverage other packages written for the React ecosystem
     plugins: [
       nodeResolve(),
+      css(),
       esbuild({
         minify: !isDev,
       }),
@@ -24,6 +28,14 @@ export default [
         // Automatically add the built script to the privacy center's static files for bundling:
         targets: [
           { src: `dist/${name}.js`, dest: "../privacy-center/public/lib/" },
+        ],
+        verbose: true,
+        hook: "writeBundle",
+      }),
+      copy({
+        // Automatically add the built css to the privacy center's static files for bundling:
+        targets: [
+          { src: `dist/${name}.css`, dest: "../privacy-center/public/lib/" },
         ],
         verbose: true,
         hook: "writeBundle",
@@ -70,7 +82,7 @@ export default [
   },
   {
     input: `src/${name}.ts`,
-    plugins: [nodeResolve(), esbuild()],
+    plugins: [nodeResolve(), css(), esbuild()],
     output: [
       {
         // Compatible with ES module imports. Apps in this repo may be able to share the code.
@@ -82,7 +94,7 @@ export default [
   },
   {
     input: `src/${name}.ts`,
-    plugins: [dts()],
+    plugins: [dts(), css()],
     output: [
       {
         file: `dist/${name}.d.ts`,